backend/requirements.txt
# FastAPI and server dependencies
fastapi
uvicorn
python-dotenv
supabase>=2.0.0
gotrue>=2.0.0

# Google Gemini AI
google-generativeai==0.3.0

# HTTP client for external APIs
httpx>=0.27.0
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utility libraries
uuid==1.30
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 